% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_snakemake_s4_snippet.R
\name{build_snakemake_s4_snippet}
\alias{build_snakemake_s4_snippet}
\title{Generate snippet for snakemake S4 object}
\usage{
build_snakemake_s4_snippet(n_input = NULL, n_output = NULL, n_param = NULL)
}
\arguments{
\item{n_input}{The number of input items to generate.}

\item{n_output}{The number of output items to generate.}

\item{n_param}{The number of parameter items to generate.}
}
\value{
A string of class glue containing the formatted snippet.
}
\description{
The function is used to produce a snippet for constructing a mock snakemake S4 object. The snippet is properly formated and ready to be pasted in the script. The user then can specify the desired input, output and params.
}
\examples{
build_snakemake_s4_snippet(n_input = 2, n_output = 2, n_param = 1)
}
\seealso{
\code{\link[=insert_snakemake_s4_snippet]{insert_snakemake_s4_snippet()}}
}
