% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_sm_s4.R
\name{create_sm_s4}
\alias{create_sm_s4}
\title{Generate Snakemake S4 snippet}
\usage{
create_sm_s4()
}
\value{
A length-one character vector containing the formatted snippet.
}
\description{
Returns the code skeleton for defining a \code{Snakemake} S4 class and a
corresponding instance. The snippet is ready to paste into an R script.
}
\examples{
cat(create_sm_s4())
}
\seealso{
\code{\link[=paste_sm_s4_code]{paste_sm_s4_code()}}
}
