% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wildcards.R
\name{expand_path}
\alias{expand_path}
\title{Replaces matching tokens in a path string.}
\usage{
expand_path(path_string, pattern, replacement)
}
\arguments{
\item{path_string}{A single character string for the path.}

\item{pattern}{A character vector of tokens to replace.}

\item{replacement}{The single string to use as a replacement.}
}
\value{
The path string with tokens replaced.
}
\description{
Replaces matching tokens in a path string.
}
