% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_fn_call.R
\name{find_fn_calls}
\alias{find_fn_calls}
\title{Find matching calls in multiple expressions}
\usage{
find_fn_calls(x, fn_name)
}
\arguments{
\item{x}{A list of expressions or calls to search.}

\item{fn_name}{A string giving the function name to locate.}
}
\value{
A list in which each element contains the calls from the corresponding
element of \code{x} that match \code{fn_name}.
}
\description{
Apply \code{find_fn_call()} to each element of a list of expressions.
}
