% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wildcards.R
\name{get_variable_path_token}
\alias{get_variable_path_token}
\title{Find which path tokens are variable.}
\usage{
get_variable_path_token(path_string, meta)
}
\arguments{
\item{path_string}{A single character string for the path.}

\item{meta}{The list of metadata tibbles.}
}
\value{
A character vector of matching tokens.
}
\description{
Find which path tokens are variable.
}
