% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mock_snakemake_project.R
\name{is_snakemake_project}
\alias{is_snakemake_project}
\title{Test if the directory is a snakemake project}
\usage{
is_snakemake_project(directory = ".")
}
\arguments{
\item{directory}{Directory to be checked. Defaults to the current working directory.}
}
\value{
Logical of length one indicating whether required files exist.
}
\description{
Searches the supplied directory for a \code{Snakefile} (either at the project
root or under \verb{workflow/}) and a configuration file (\code{config.yaml} or
\code{config.yml} under \verb{config/}). Both must be present for the function to
return \code{TRUE}.
}
\examples{
\dontrun{
demo_dir <- mock_snakemake_project(tempfile("snakemake-demo-"))
is_snakemake_project(demo_dir)
}
}
