% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wildcards.R
\name{read_project_meta}
\alias{read_project_meta}
\title{Reads all project metadata files from the config/ directory.}
\usage{
read_project_meta()
}
\value{
A list of tibbles, one for each .csv file found.
}
\description{
Reads all project metadata files from the config/ directory.
}
