% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_fn_call.R
\name{recurse_compact}
\alias{recurse_compact}
\title{Recursively compact a list}
\usage{
recurse_compact(x)
}
\arguments{
\item{x}{A list, possibly containing nested lists.}
}
\value{
A list with \code{NULL} or empty list elements removed.
}
\description{
Apply \code{purrr::compact()} recursively to drop \code{NULL} or empty list elements from a list.
}
\keyword{internal}
