% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_fn_call.R
\name{recurse_flatten}
\alias{recurse_flatten}
\title{Recursively flatten a list}
\usage{
recurse_flatten(x)
}
\arguments{
\item{x}{A list, possibly containing nested lists.}
}
\value{
A list with flattened hierarchy.
}
\description{
Apply \code{purrr::list_flatten()} recursively to reduce the depth of a list as much as possible.
}
