% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wildcards.R
\name{which_path_expendable}
\alias{which_path_expendable}
\title{Finds which parts (tokens) of a path string are present in the metadata.}
\usage{
which_path_expendable(path_string, meta)
}
\arguments{
\item{path_string}{A single character string for the path.}

\item{meta}{The list of metadata tibbles.}
}
\value{
A character vector of matching tokens.
}
\description{
Finds which parts (tokens) of a path string are present in the metadata.
}
