% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rscript_to_snakemake_rule.R
\name{build_rule}
\alias{build_rule}
\title{Build a Snakemake rule}
\usage{
build_rule(output, script, rule_name = NULL, input = NULL, params = NULL)
}
\arguments{
\item{script}{Path to the script.}

\item{rule_name}{Explicit rule name. Defaults to the script file name without extension.}

\item{input, output}{Named lists pointing to input and output files.}

\item{params}{Named list of parameter.}
}
\value{
A single character string containing the formatted rule.
}
\description{
Convenience helper for writing a Snakemake rule from its components.
}
