% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_snakemake_s4_snippet.R
\name{build_snakemake_s4_snippet}
\alias{build_snakemake_s4_snippet}
\title{Generate Snakemake S4 snippet}
\usage{
build_snakemake_s4_snippet()
}
\value{
A length-one character vector containing the formatted snippet.
}
\description{
Returns the code skeleton for defining a \code{Snakemake} S4 class and a
corresponding instance. The snippet is ready to paste into an R script.
}
\examples{
cat(build_snakemake_s4_snippet())
}
\seealso{
\code{\link[=insert_snakemake_s4_snippet]{insert_snakemake_s4_snippet()}}
}
