% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_snakemake_s4_snippet.R
\name{build_snakemake_s4_snippet}
\alias{build_snakemake_s4_snippet}
\title{Generate snippet for a Snakemake S4 object}
\usage{
build_snakemake_s4_snippet(n_input = 1, n_output = 1, n_param = 1)
}
\arguments{
\item{n_input, n_output, n_param}{Single non-negative integers indicating how
many placeholder items to generate for each slot.}
}
\value{
A \code{\link[glue:glue]{glue::glue}} string containing the formatted snippet.
}
\description{
Build a formatted snippet for a minimal Snakemake S4 class definition and
instance. You can control how many placeholder entries are generated for the
\code{input}, \code{output}, and \code{params} slots.
}
\examples{
build_snakemake_s4_snippet(n_input = 2, n_output = 2, n_param = 1)
}
\seealso{
\code{\link[=insert_snakemake_s4_snippet]{insert_snakemake_s4_snippet()}}
}
