% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_fn_call.R
\name{find_fn_call}
\alias{find_fn_call}
\title{Find matching calls within an expression}
\usage{
find_fn_call(x, fn_name)
}
\arguments{
\item{x}{An expression to search.}

\item{fn_name}{A string giving the function name to locate.}
}
\value{
A list of calls matching \code{fn_name}. Returns an empty list if no matches are found.
}
\description{
Recursively walks through an expression and collects calls whose function name matches
\code{fn_name}.
}
