% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert_snakemake_s4_snippet.R
\name{insert_snakemake_s4_snippet}
\alias{insert_snakemake_s4_snippet}
\title{Paste Snakemake S4 snippet into an RStudio script}
\usage{
insert_snakemake_s4_snippet()
}
\value{
Invisibly returns the inserted snippet as a character vector.
}
\description{
Inserts the output of \code{\link[=build_snakemake_s4_snippet]{build_snakemake_s4_snippet()}} at the current cursor position in the
active RStudio source editor.
}
\examples{
if (rstudioapi::isAvailable()) {
  paste_sm_s4_code()
}
}
\seealso{
\code{\link[=build_snakemake_s4_snippet]{build_snakemake_s4_snippet()}}
}
