% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert_snakemake_s4_snippet.R
\name{insert_snakemake_s4_snippet}
\alias{insert_snakemake_s4_snippet}
\title{Paste snippet for snakemake S4 object}
\usage{
insert_snakemake_s4_snippet(...)
}
\value{
Invisibly returns the inserted snippet as a character vector.
}
\description{
Paste the snippet for constructing snakemake S4 object. The snippet is generated by \code{\link[=build_snakemake_s4_snippet]{build_snakemake_s4_snippet()}}.
}
\examples{
if (rstudioapi::isAvailable()) {
  paste_sm_s4_code()
}
}
\seealso{
\code{\link[=build_snakemake_s4_snippet]{build_snakemake_s4_snippet()}}
}
