% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mock_snakemake_project.R
\name{is_snakemake_project}
\alias{is_snakemake_project}
\title{Test if the directory is a snakemake project}
\usage{
is_snakemake_project(directory = ".")
}
\arguments{
\item{directory}{directory to be checked}
}
\value{
logical of length 1
}
\description{
Searches the directory for the presence of Snakefile in root or in workflow.
}
